requirements.txt
streamlit
sentence-transformers
pdfminer.six
spacy
hnswlib
pandas
numpy
pyngrok
scikit-learn
tqdm
requests
